feat(chat): widget Home hero, header color, thread-only expand, icon refactor
- Add `headerColor` column (drop `expandedWidthPx`); fixed 640px expanded width
- Home view paints a hero band in `headerColor` with auto-contrast text and
  inline close button; standard frame header is hidden at Home root
- Expand mode is thread-only; sticky preference restored on entering a thread
- Settings: add Header Color picker; drop expanded-width slider; preview
  popup forces a fresh load via `?v=` + `data-v` cache-bust through to the
  widget bundle and `/api/chat/config`
- Extract icon registry to `@auxx/ui/components/icon-data`; KB list and
  article cards render Lucide icons by id (stored in `Article.emoji`)
