+++ packages/database/drizzle/0170_chat_widget_header_color_drop_expanded_width.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "ChatWidget" ADD COLUMN "headerColor" text DEFAULT '#1F2329' NOT NULL;--> statement-breakpoint
+ALTER TABLE "ChatWidget" DROP COLUMN "expandedWidthPx";
