+++ packages/database/src/db/schema/chat-widget.ts
@@ -6,7 +6,6 @@ import {
   type AnyPgColumn,
   boolean,
   index,
-  integer,
   jsonb,
   pgTable,
   text,
@@ -39,6 +38,10 @@ export const ChatWidget = pgTable(
     title: text().default('Chat Support').notNull(),
     subtitle: text(),
     primaryColor: text().default('#4F46E5').notNull(),
+    /** Hex color for the Home hero band. Independent of `primaryColor` so the
+     * launcher and the header can be tuned separately. Text color is
+     * auto-derived client-side from luminance. */
+    headerColor: text().default('#1F2329').notNull(),
     logoLight: text(),
     logoDark: text(),
     logoLightId: text().references((): AnyPgColumn => MediaAsset.id, { onUpdate: 'cascade' }),
@@ -65,7 +68,6 @@ export const ChatWidget = pgTable(
     homeShowRecentMessage: boolean().default(true).notNull(),
     homeShowSendMessageCta: boolean().default(true).notNull(),
     brandingFooterEnabled: boolean().default(true).notNull(),
-    expandedWidthPx: integer().default(720).notNull(),
 
     // v2 KB linkage (one widget → at most one KB)
     knowledgeBaseId: text().references((): AnyPgColumn => KnowledgeBase.id, {
