+++ packages/ui/package.json
@@ -23,6 +23,7 @@
     "./components/menu-styles": "./src/components/menu-styles.ts",
     "./components/*": "./src/components/*.tsx",
     "./components/emojis": "./src/components/emojis.ts",
+    "./components/icon-data": "./src/components/icon-data.ts",
     "./components/icons/*": "./src/components/icons/*.tsx",
     "./hooks/*": "./src/hooks/*.ts",
     "./hooks": "./src/hooks/index.ts",
