+++ packages/ui/src/components/icons.tsx
@@ -3,218 +3,13 @@
 
 import { cn } from '@auxx/ui/lib/utils'
 import { cva, type VariantProps } from 'class-variance-authority'
-import type { LucideIcon } from 'lucide-react'
-import {
-  Activity,
-  // Status
-  AlertCircle,
-  AlertTriangle,
-  AlignCenter,
-  Archive,
-  ArrowDown,
-  ArrowLeft,
-  ArrowLeftRight,
-  ArrowRight,
-  ArrowUp,
-  ArrowUpDown,
-  AtSign,
-  Award,
-  Ban,
-  // Data/Charts
-  BarChart,
-  Battery,
-  Bell,
-  BellRing,
-  Bookmark,
-  BookOpen,
-  Box,
-  Boxes,
-  Braces,
-  Brain,
-  // Business
-  Briefcase,
-  Building,
-  Building2,
-  Calculator,
-  Calendar,
-  CalendarClock,
-  Camera,
-  CaseSensitive,
-  Check,
-  CheckCircle,
-  CheckSquare,
-  ChevronDown,
-  ChevronLeft,
-  ChevronRight,
-  ChevronsUpDown,
-  ChevronUp,
-  Circle,
-  CircleUser,
-  Clipboard,
-  ClipboardCheck,
-  Clock,
-  Cloud,
-  Code,
-  Columns3,
-  // Navigation
-  Compass,
-  Copy,
-  CreditCard,
-  Crosshair,
-  Crown,
-  Database,
-  DollarSign,
-  Download,
-  // Actions
-  Edit,
-  Equal,
-  ExternalLink,
-  // Additional useful icons
-  Eye,
-  EyeOff,
-  // Files/Documents
-  File,
-  FileAudio,
-  FileCode,
-  FileImage,
-  FileJson,
-  FileSpreadsheet,
-  FileText,
-  FileType,
-  FileVideo,
-  Filter,
-  Fingerprint,
-  Flag,
-  Folder,
-  FolderKanban,
-  FolderOpen,
-  Gift,
-  // Workflow-specific icons
-  GitBranch,
-  Globe,
-  Grid,
-  Hammer,
-  Handshake,
-  Hash,
-  // Text formatting
-  Heading1,
-  Heading2,
-  Heading3,
-  Heart,
-  HelpCircle,
-  Hexagon,
-  // General UI
-  History,
-  Home,
-  // Media
-  Image,
-  Inbox,
-  Info,
-  Key,
-  Layers,
-  Layout,
-  LayoutDashboard,
-  Lightbulb,
-  LineChart,
-  Link,
-  Link2,
-  List,
-  ListChecks,
-  ListFilter,
-  ListOrdered,
-  Lock,
-  // Communication
-  Mail,
-  Mails,
-  Map,
-  MapPin,
-  Megaphone,
-  Menu,
-  MessageCircle,
-  MessageSquare,
-  Minus,
-  Moon,
-  MoreHorizontal,
-  MoreVertical,
-  Music,
-  Navigation,
-  Octagon,
-  Package,
-  Paperclip,
-  Pause,
-  PenTool,
-  Percent,
-  Phone,
-  PieChart,
-  Play,
-  Plug,
-  Plus,
-  Power,
-  Printer,
-  QrCode,
-  Quote,
-  Receipt,
-  RefreshCw,
-  Regex,
-  Repeat,
-  Replace,
-  RotateCcw,
-  Save,
-  Scan,
-  Scissors,
-  Search,
-  Send,
-  Server,
-  Settings,
-  Share,
-  Share2,
-  Shield,
-  ShieldCheck,
-  ShoppingCart,
-  Sparkles,
-  Square,
-  // Misc
-  Star,
-  StickyNote,
-  Sun,
-  Table,
-  Tag,
-  Tags,
-  Target,
-  Terminal,
-  TextCursorInput,
-  Ticket,
-  ToggleLeft,
-  Trash,
-  TrendingDown,
-  TrendingUp,
-  Triangle,
-  Truck,
-  Type,
-  Undo2,
-  Unlock,
-  Upload,
-  // User/Profile
-  User,
-  UserCheck,
-  UserMinus,
-  UserPlus,
-  Users,
-  UserX,
-  Variable,
-  Video,
-  Volume2,
-  VolumeX,
-  Warehouse,
-  Webhook,
-  Wifi,
-  // Tools
-  Wrench,
-  X,
-  XCircle,
-  Zap,
-} from 'lucide-react'
 import type React from 'react'
+import { getIcon, ICON_DATA, type IconItem } from './icon-data'
+
+// Re-export from the leaf module so non-React callers (chat-widget bundle)
+// can import the id → Lucide-component catalog without pulling in this file's
+// `'use client'` + CVA + EntityIcon.
+export { getIcon, ICON_DATA, type IconItem }
 
 /** Color configuration for icons */
 export interface IconColor {
@@ -352,233 +147,6 @@ export const DEFAULT_COLOR = 'gray'
 /** Get color configuration by ID */
 export const getIconColor = (colorId: string): IconColor =>
   ICON_COLORS.find((c) => c.id === colorId) ?? ICON_COLORS[0]!
-
-/** Icon item definition */
-export interface IconItem {
-  /** Unique identifier (icon name in kebab-case) */
-  id: string
-  /** Display name */
-  label: string
-  /** Lucide icon component */
-  icon: LucideIcon
-}
-
-/** Curated list of commonly used icons */
-export const ICON_DATA: IconItem[] = [
-  // General UI
-  { id: 'history', label: 'History', icon: History },
-  { id: 'home', label: 'Home', icon: Home },
-  { id: 'settings', label: 'Settings', icon: Settings },
-  { id: 'search', label: 'Search', icon: Search },
-  { id: 'menu', label: 'Menu', icon: Menu },
-  { id: 'x', label: 'Close', icon: X },
-  { id: 'plus', label: 'Plus', icon: Plus },
-  { id: 'minus', label: 'Minus', icon: Minus },
-  { id: 'equal', label: 'Equal', icon: Equal },
-  { id: 'check', label: 'Check', icon: Check },
-  { id: 'chevron-down', label: 'Chevron Down', icon: ChevronDown },
-  { id: 'chevron-up', label: 'Chevron Up', icon: ChevronUp },
-  { id: 'chevron-left', label: 'Chevron Left', icon: ChevronLeft },
-  { id: 'chevron-right', label: 'Chevron Right', icon: ChevronRight },
-  { id: 'chevrons-up-down', label: 'Chevrons Up Down', icon: ChevronsUpDown },
-  { id: 'arrow-left', label: 'Arrow Left', icon: ArrowLeft },
-  { id: 'arrow-right', label: 'Arrow Right', icon: ArrowRight },
-  { id: 'arrow-up', label: 'Arrow Up', icon: ArrowUp },
-  { id: 'arrow-down', label: 'Arrow Down', icon: ArrowDown },
-  { id: 'arrow-left-right', label: 'Arrow Left Right', icon: ArrowLeftRight },
-  // Communication
-  { id: 'mail', label: 'Mail', icon: Mail },
-  { id: 'message-square', label: 'Message Square', icon: MessageSquare },
-  { id: 'message-circle', label: 'Message Circle', icon: MessageCircle },
-  { id: 'phone', label: 'Phone', icon: Phone },
-  { id: 'bell', label: 'Bell', icon: Bell },
-  { id: 'bell-ring', label: 'Bell Ring', icon: BellRing },
-  { id: 'send', label: 'Send', icon: Send },
-  { id: 'inbox', label: 'Inbox', icon: Inbox },
-  { id: 'mails', label: 'Mails', icon: Mails },
-  { id: 'at-sign', label: 'At Sign', icon: AtSign },
-  // User/Profile
-  { id: 'user', label: 'User', icon: User },
-  { id: 'users', label: 'Users', icon: Users },
-  { id: 'user-plus', label: 'User Plus', icon: UserPlus },
-  { id: 'user-minus', label: 'User Minus', icon: UserMinus },
-  { id: 'user-check', label: 'User Check', icon: UserCheck },
-  { id: 'user-x', label: 'User X', icon: UserX },
-  { id: 'circle-user', label: 'Circle User', icon: CircleUser },
-  // Files/Documents
-  { id: 'file', label: 'File', icon: File },
-  { id: 'file-text', label: 'File Text', icon: FileText },
-  { id: 'file-image', label: 'File Image', icon: FileImage },
-  { id: 'file-video', label: 'File Video', icon: FileVideo },
-  { id: 'file-audio', label: 'File Audio', icon: FileAudio },
-  { id: 'file-type', label: 'File Type', icon: FileType },
-  { id: 'file-spreadsheet', label: 'File Spreadsheet', icon: FileSpreadsheet },
-  { id: 'file-code', label: 'File Code', icon: FileCode },
-  { id: 'file-json', label: 'File JSON', icon: FileJson },
-  { id: 'folder-kanban', label: 'Folder Kanban', icon: FolderKanban },
-  { id: 'archive', label: 'Archive', icon: Archive },
-  { id: 'folder', label: 'Folder', icon: Folder },
-  { id: 'folder-open', label: 'Folder Open', icon: FolderOpen },
-  { id: 'upload', label: 'Upload', icon: Upload },
-  { id: 'download', label: 'Download', icon: Download },
-  { id: 'paperclip', label: 'Paperclip', icon: Paperclip },
-  { id: 'link', label: 'Link', icon: Link },
-  // Actions
-  { id: 'edit', label: 'Edit', icon: Edit },
-  { id: 'trash', label: 'Trash', icon: Trash },
-  { id: 'copy', label: 'Copy', icon: Copy },
-  { id: 'clipboard', label: 'Clipboard', icon: Clipboard },
-  { id: 'clipboard-check', label: 'Clipboard Check', icon: ClipboardCheck },
-  { id: 'share-2', label: 'Share 2', icon: Share2 },
-  { id: 'save', label: 'Save', icon: Save },
-  { id: 'share', label: 'Share', icon: Share },
-  { id: 'external-link', label: 'External Link', icon: ExternalLink },
-  { id: 'more-horizontal', label: 'More Horizontal', icon: MoreHorizontal },
-  { id: 'more-vertical', label: 'More Vertical', icon: MoreVertical },
-  // Media
-  { id: 'image', label: 'Image', icon: Image },
-  { id: 'video', label: 'Video', icon: Video },
-  { id: 'music', label: 'Music', icon: Music },
-  { id: 'camera', label: 'Camera', icon: Camera },
-  { id: 'play', label: 'Play', icon: Play },
-  { id: 'pause', label: 'Pause', icon: Pause },
-  { id: 'volume', label: 'Volume', icon: Volume2 },
-  { id: 'volume-x', label: 'Volume Off', icon: VolumeX },
-  // Business
-  { id: 'briefcase', label: 'Briefcase', icon: Briefcase },
-  { id: 'building', label: 'Building', icon: Building },
-  { id: 'building-2', label: 'Building 2', icon: Building2 },
-  { id: 'calendar', label: 'Calendar', icon: Calendar },
-  { id: 'clock', label: 'Clock', icon: Clock },
-  { id: 'credit-card', label: 'Credit Card', icon: CreditCard },
-  { id: 'dollar-sign', label: 'Dollar Sign', icon: DollarSign },
-  { id: 'shopping-cart', label: 'Shopping Cart', icon: ShoppingCart },
-  { id: 'package', label: 'Package', icon: Package },
-  { id: 'truck', label: 'Truck', icon: Truck },
-  { id: 'megaphone', label: 'Megaphone', icon: Megaphone },
-  { id: 'warehouse', label: 'Warehouse', icon: Warehouse },
-  { id: 'receipt', label: 'Receipt', icon: Receipt },
-  { id: 'handshake', label: 'Handshake', icon: Handshake },
-  { id: 'percent', label: 'Percent', icon: Percent },
-  { id: 'crown', label: 'Crown', icon: Crown },
-  // Data/Charts
-  { id: 'bar-chart', label: 'Bar Chart', icon: BarChart },
-  { id: 'line-chart', label: 'Line Chart', icon: LineChart },
-  { id: 'pie-chart', label: 'Pie Chart', icon: PieChart },
-  { id: 'trending-up', label: 'Trending Up', icon: TrendingUp },
-  { id: 'trending-down', label: 'Trending Down', icon: TrendingDown },
-  { id: 'activity', label: 'Activity', icon: Activity },
-  // Status
-  { id: 'alert-circle', label: 'Alert Circle', icon: AlertCircle },
-  { id: 'alert-triangle', label: 'Alert Triangle', icon: AlertTriangle },
-  { id: 'info', label: 'Info', icon: Info },
-  { id: 'help-circle', label: 'Help Circle', icon: HelpCircle },
-  { id: 'check-circle', label: 'Check Circle', icon: CheckCircle },
-  { id: 'x-circle', label: 'X Circle', icon: XCircle },
-  { id: 'ban', label: 'Ban', icon: Ban },
-  // Tools
-  { id: 'wrench', label: 'Wrench', icon: Wrench },
-  { id: 'hammer', label: 'Hammer', icon: Hammer },
-  { id: 'key', label: 'Key', icon: Key },
-  { id: 'lock', label: 'Lock', icon: Lock },
-  { id: 'unlock', label: 'Unlock', icon: Unlock },
-  { id: 'shield', label: 'Shield', icon: Shield },
-  { id: 'shield-check', label: 'Shield Check', icon: ShieldCheck },
-  { id: 'zap', label: 'Zap', icon: Zap },
-  { id: 'plug', label: 'Plug', icon: Plug },
-  { id: 'lightbulb', label: 'Lightbulb', icon: Lightbulb },
-  // Navigation
-  { id: 'compass', label: 'Compass', icon: Compass },
-  { id: 'map', label: 'Map', icon: Map },
-  { id: 'map-pin', label: 'Map Pin', icon: MapPin },
-  { id: 'globe', label: 'Globe', icon: Globe },
-  { id: 'navigation', label: 'Navigation', icon: Navigation },
-  // Misc
-  { id: 'star', label: 'Star', icon: Star },
-  { id: 'heart', label: 'Heart', icon: Heart },
-  { id: 'bookmark', label: 'Bookmark', icon: Bookmark },
-  { id: 'tag', label: 'Tag', icon: Tag },
-  { id: 'tags', label: 'Tags', icon: Tags },
-  { id: 'flag', label: 'Flag', icon: Flag },
-  { id: 'award', label: 'Award', icon: Award },
-  { id: 'gift', label: 'Gift', icon: Gift },
-  { id: 'sparkles', label: 'Sparkles', icon: Sparkles },
-  // Additional useful icons
-  { id: 'eye', label: 'Eye', icon: Eye },
-  { id: 'eye-off', label: 'Eye Off', icon: EyeOff },
-  { id: 'filter', label: 'Filter', icon: Filter },
-  { id: 'refresh', label: 'Refresh', icon: RefreshCw },
-  { id: 'rotate', label: 'Rotate', icon: RotateCcw },
-  { id: 'layers', label: 'Layers', icon: Layers },
-  { id: 'grid', label: 'Grid', icon: Grid },
-  { id: 'list', label: 'List', icon: List },
-  { id: 'table', label: 'Table', icon: Table },
-  { id: 'columns', label: 'Columns', icon: Columns3 },
-  { id: 'layout', label: 'Layout', icon: Layout },
-  { id: 'layout-dashboard', label: 'Layout Dashboard', icon: LayoutDashboard },
-  { id: 'book-open', label: 'Book Open', icon: BookOpen },
-  { id: 'code', label: 'Code', icon: Code },
-  { id: 'terminal', label: 'Terminal', icon: Terminal },
-  { id: 'database', label: 'Database', icon: Database },
-  { id: 'server', label: 'Server', icon: Server },
-  { id: 'cloud', label: 'Cloud', icon: Cloud },
-  { id: 'wifi', label: 'Wifi', icon: Wifi },
-  { id: 'battery', label: 'Battery', icon: Battery },
-  { id: 'power', label: 'Power', icon: Power },
-  { id: 'sun', label: 'Sun', icon: Sun },
-  { id: 'moon', label: 'Moon', icon: Moon },
-  { id: 'circle', label: 'Circle', icon: Circle },
-  { id: 'square', label: 'Square', icon: Square },
-  { id: 'triangle', label: 'Triangle', icon: Triangle },
-  { id: 'hexagon', label: 'Hexagon', icon: Hexagon },
-  { id: 'target', label: 'Target', icon: Target },
-  { id: 'crosshair', label: 'Crosshair', icon: Crosshair },
-  { id: 'fingerprint', label: 'Fingerprint', icon: Fingerprint },
-  { id: 'qr-code', label: 'QR Code', icon: QrCode },
-  { id: 'printer', label: 'Printer', icon: Printer },
-  { id: 'scan', label: 'Scan', icon: Scan },
-  // Text formatting
-  { id: 'heading-1', label: 'Heading 1', icon: Heading1 },
-  { id: 'heading-2', label: 'Heading 2', icon: Heading2 },
-  { id: 'heading-3', label: 'Heading 3', icon: Heading3 },
-  { id: 'list-ordered', label: 'Numbered List', icon: ListOrdered },
-  { id: 'quote', label: 'Quote', icon: Quote },
-  // Workflow-specific icons
-  { id: 'git-branch', label: 'Git Branch', icon: GitBranch },
-  { id: 'sticky-note', label: 'Sticky Note', icon: StickyNote },
-  { id: 'variable', label: 'Variable', icon: Variable },
-  { id: 'brain', label: 'Brain', icon: Brain },
-  { id: 'octagon', label: 'Octagon', icon: Octagon },
-  { id: 'repeat', label: 'Repeat', icon: Repeat },
-  { id: 'undo-2', label: 'Undo', icon: Undo2 },
-  { id: 'arrows-up-down', label: 'Arrows Up Down', icon: ArrowUpDown },
-  { id: 'webhook', label: 'Webhook', icon: Webhook },
-  { id: 'ticket', label: 'Ticket', icon: Ticket },
-  { id: 'box', label: 'Box', icon: Box },
-  { id: 'boxes', label: 'Boxes', icon: Boxes },
-  { id: 'type', label: 'Type', icon: Type },
-  { id: 'hash', label: 'Hash', icon: Hash },
-  { id: 'toggle-left', label: 'Toggle Left', icon: ToggleLeft },
-  { id: 'braces', label: 'Braces', icon: Braces },
-  { id: 'pen-tool', label: 'Pen Tool', icon: PenTool },
-  { id: 'scissors', label: 'Scissors', icon: Scissors },
-  { id: 'text', label: 'Text', icon: CaseSensitive },
-  { id: 'text-cursor-input', label: 'Text Cursor Input', icon: TextCursorInput },
-  { id: 'list-filter', label: 'List Filter', icon: ListFilter },
-  { id: 'calendar-clock', label: 'Calendar Clock', icon: CalendarClock },
-  { id: 'list-checks', label: 'List Checks', icon: ListChecks },
-  { id: 'check-square', label: 'Check Square', icon: CheckSquare },
-  { id: 'link-2', label: 'Link 2', icon: Link2 },
-  { id: 'calculator', label: 'Calculator', icon: Calculator },
-  { id: 'case-sensitive', label: 'Case Sensitive', icon: CaseSensitive },
-  { id: 'align-center', label: 'Align Center', icon: AlignCenter },
-  { id: 'replace', label: 'Replace', icon: Replace },
-  { id: 'regex', label: 'Regex', icon: Regex },
-]
-
-/** Get icon item by ID */
-export const getIcon = (iconId: string): IconItem | undefined =>
-  ICON_DATA.find((item) => item.id === iconId)
 
 /** EntityIcon variants using CVA */
 export const entityIconVariants = cva('flex items-center justify-center shrink-0', {
